requirements.txt
Django==5.0.7
psycopg2==2.9.9
djangorestframework==3.15.2
Pygments==2.18.0